Extract MockHTTPClient
